Move TES and Cosmos to separate project to prepare for future CosmosDB access by TriggerService.
