+++ CromwellOnAzure.sln
@@ -23,9 +23,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TesApi.Tests", "src\TesApi.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TriggerService", "src\TriggerService\TriggerService.csproj", "{BB61BF85-0DE4-469B-85D4-3744E1B03F89}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "src\Common\Common.csproj", "{63C91E1C-640C-4281-8388-971908009CAE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tes", "src\Tes\Tes.csproj", "{63C91E1C-640C-4281-8388-971908009CAE}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TriggerService.Tests", "src\TriggerService.Tests\TriggerService.Tests.csproj", "{6AE9D93D-FBFA-4B94-8046-CDF10886F15B}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "src\Common\Common.csproj", "{1D5F3C4A-5A69-4A86-BF80-D075DBDD0307}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -65,6 +67,10 @@ Global
 		{6AE9D93D-FBFA-4B94-8046-CDF10886F15B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6AE9D93D-FBFA-4B94-8046-CDF10886F15B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6AE9D93D-FBFA-4B94-8046-CDF10886F15B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D5F3C4A-5A69-4A86-BF80-D075DBDD0307}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D5F3C4A-5A69-4A86-BF80-D075DBDD0307}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D5F3C4A-5A69-4A86-BF80-D075DBDD0307}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D5F3C4A-5A69-4A86-BF80-D075DBDD0307}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
