+++ src/Tes/Models/TesCreateTaskResponse.cs
@@ -1,0 +1,136 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/*
+ * Task Execution Service
+ *
+ * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+ *
+ * OpenAPI spec version: 0.3.0
+ * 
+ * Generated by: https://openapi-generator.tech
+ */
+
+using System;
+using System.Runtime.Serialization;
+using System.Text;
+using Newtonsoft.Json;
+
+namespace Tes.Models
+{
+    /// <summary>
+    /// CreateTaskResponse describes a response from the CreateTask endpoint.
+    /// </summary>
+    [DataContract]
+    public partial class TesCreateTaskResponse : IEquatable<TesCreateTaskResponse>
+    {
+        /// <summary>
+        /// Task identifier assigned by the server.
+        /// </summary>
+        /// <value>Task identifier assigned by the server.</value>
+        [DataMember(Name = "id")]
+        public string Id { get; set; }
+
+        /// <summary>
+        /// Returns the string presentation of the object
+        /// </summary>
+        /// <returns>String presentation of the object</returns>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class TesCreateTaskResponse {\n");
+            sb.Append("  Id: ").Append(Id).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns the JSON string presentation of the object
+        /// </summary>
+        /// <returns>JSON string presentation of the object</returns>
+        public string ToJson()
+        {
+            return JsonConvert.SerializeObject(this, Formatting.Indented);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        /// <param name="obj">Object to be compared</param>
+        /// <returns>Boolean</returns>
+        public override bool Equals(object obj)
+        {
+            if (obj is null)
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, obj))
+            {
+                return true;
+            }
+
+            return obj.GetType() == GetType() && Equals((TesCreateTaskResponse)obj);
+        }
+
+        /// <summary>
+        /// Returns true if TesCreateTaskResponse instances are equal
+        /// </summary>
+        /// <param name="other">Instance of TesCreateTaskResponse to be compared</param>
+        /// <returns>Boolean</returns>
+        public bool Equals(TesCreateTaskResponse other)
+        {
+            if (other is null)
+            {
+                return false;
+            }
+
+            if (ReferenceEquals(this, other))
+            {
+                return true;
+            }
+
+            return
+                (
+                    Id == other.Id ||
+                    Id != null &&
+                    Id.Equals(other.Id)
+                );
+        }
+
+        /// <summary>
+        /// Gets the hash code
+        /// </summary>
+        /// <returns>Hash code</returns>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                var hashCode = 41;
+                // Suitable nullity checks etc, of course :)
+                if (Id != null)
+                {
+                    hashCode = hashCode * 59 + Id.GetHashCode();
+                }
+
+                return hashCode;
+            }
+        }
+
+        #region Operators
+#pragma warning disable 1591
+
+        public static bool operator ==(TesCreateTaskResponse left, TesCreateTaskResponse right)
+        {
+            return Equals(left, right);
+        }
+
+        public static bool operator !=(TesCreateTaskResponse left, TesCreateTaskResponse right)
+        {
+            return !Equals(left, right);
+        }
+
+#pragma warning restore 1591
+        #endregion Operators
+    }
+}
