+++ src/Tes/Models/TesExecutorLog.cs
@@ -16,7 +16,7 @@ using System.Runtime.Serialization;
 using System.Text;
 using Newtonsoft.Json;
 
-namespace TesApi.Models
+namespace Tes.Models
 {
     /// <summary>
     /// ExecutorLog describes logging information related to an Executor.
