+++ src/Tes/Models/TesFileType.cs
@@ -14,7 +14,7 @@
 using System.Runtime.Serialization;
 using Newtonsoft.Json;
 
-namespace TesApi.Models
+namespace Tes.Models
 {
     /// <summary>
     /// Gets or Sets tesFileType
