+++ src/Tes/Models/TesTask.cs
@@ -18,7 +18,7 @@ using System.Runtime.Serialization;
 using System.Text;
 using Newtonsoft.Json;
 
-namespace TesApi.Models
+namespace Tes.Models
 {
     /// <summary>
     /// Task describes an instance of a task.
