+++ src/TesApi.Web/AzureProxy.cs
@@ -27,7 +27,7 @@ using Microsoft.Rest;
 using Microsoft.WindowsAzure.Storage.Blob;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using TesApi.Models;
+using Tes.Models;
 
 namespace TesApi.Web
 {
@@ -729,11 +729,11 @@ namespace TesApi.Web
         /// Gets an authenticated Azure Client instance
         /// </summary>
         /// <returns>An authenticated Azure Client instance</returns>
-        private static async Task<Azure.IAuthenticated> GetAzureManagementClientAsync()
+        private static async Task<Microsoft.Azure.Management.Fluent.Azure.IAuthenticated> GetAzureManagementClientAsync()
         {
             var accessToken = await GetAzureAccessTokenAsync();
             var azureCredentials = new AzureCredentials(new TokenCredentials(accessToken), null, null, AzureEnvironment.AzureGlobalCloud);
-            var azureClient = Azure.Authenticate(azureCredentials);
+            var azureClient = Microsoft.Azure.Management.Fluent.Azure.Authenticate(azureCredentials);
 
             return azureClient;
         }
