+++ src/TesApi.Web/ContractResolvers/FullTesTaskContractResolver.cs
@@ -6,7 +6,8 @@ using System.Collections.Generic;
 using System.Reflection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using TesApi.Models;
+using Tes.Models;
+using Tes.Repository;
 
 namespace TesApi.Web
 {
