+++ src/TesApi.Web/ContractResolvers/MinimalTesTaskContractResolver.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using TesApi.Models;
+using Tes.Models;
 
 namespace TesApi.Web
 {
