+++ src/TesApi.Web/DeleteCompletedBatchJobsHostedService.cs
@@ -4,7 +4,8 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using TesApi.Models;
+using Tes.Models;
+using Tes.Repository;
 
 namespace TesApi.Web
 {
