--- src/TesApi.Web/Models/TesInput.cs
@@ -1,218 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-/*
- * Task Execution Service
- *
- * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
- *
- * OpenAPI spec version: 0.3.0
- * 
- * Generated by: https://openapi-generator.tech
- */
-
-using System;
-using System.Runtime.Serialization;
-using System.Text;
-using Newtonsoft.Json;
-
-namespace TesApi.Models
-{
-    /// <summary>
-    /// Input describes Task input files.
-    /// </summary>
-    [DataContract]
-    public partial class TesInput : IEquatable<TesInput>
-    {
-        /// <summary>
-        /// Gets or Sets Name
-        /// </summary>
-        [DataMember(Name = "name")]
-        public string Name { get; set; }
-
-        /// <summary>
-        /// Gets or Sets Description
-        /// </summary>
-        [DataMember(Name = "description")]
-        public string Description { get; set; }
-
-        /// <summary>
-        /// REQUIRED, unless \&quot;content\&quot; is set.  URL in long term storage, for example: s3://my-object-store/file1 gs://my-bucket/file2 file:///path/to/my/file /path/to/my/file etc...
-        /// </summary>
-        /// <value>REQUIRED, unless \&quot;content\&quot; is set.  URL in long term storage, for example: s3://my-object-store/file1 gs://my-bucket/file2 file:///path/to/my/file /path/to/my/file etc...</value>
-        [DataMember(Name = "url")]
-        public string Url { get; set; }
-
-        /// <summary>
-        /// Path of the file inside the container. Must be an absolute path.
-        /// </summary>
-        /// <value>Path of the file inside the container. Must be an absolute path.</value>
-        [DataMember(Name = "path")]
-        public string Path { get; set; }
-
-        /// <summary>
-        /// Gets or Sets Type
-        /// </summary>
-        [DataMember(Name = "type")]
-        public TesFileType Type { get; set; }
-
-        /// <summary>
-        /// File content literal.  Implementations should support a minimum of 128 KiB in this field and may define its own maximum. UTF-8 encoded  If content is not empty, \&quot;url\&quot; must be ignored.
-        /// </summary>
-        /// <value>File content literal.  Implementations should support a minimum of 128 KiB in this field and may define its own maximum. UTF-8 encoded  If content is not empty, \&quot;url\&quot; must be ignored.</value>
-        [DataMember(Name = "content")]
-        public string Content { get; set; }
-
-        /// <summary>
-        /// Returns the string presentation of the object
-        /// </summary>
-        /// <returns>String presentation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder();
-            sb.Append("class TesInput {\n");
-            sb.Append("  Name: ").Append(Name).Append("\n");
-            sb.Append("  Description: ").Append(Description).Append("\n");
-            sb.Append("  Url: ").Append(Url).Append("\n");
-            sb.Append("  Path: ").Append(Path).Append("\n");
-            sb.Append("  Type: ").Append(Type).Append("\n");
-            sb.Append("  Content: ").Append(Content).Append("\n");
-            sb.Append("}\n");
-            return sb.ToString();
-        }
-
-        /// <summary>
-        /// Returns the JSON string presentation of the object
-        /// </summary>
-        /// <returns>JSON string presentation of the object</returns>
-        public string ToJson()
-        {
-            return JsonConvert.SerializeObject(this, Formatting.Indented);
-        }
-
-        /// <summary>
-        /// Returns true if objects are equal
-        /// </summary>
-        /// <param name="obj">Object to be compared</param>
-        /// <returns>Boolean</returns>
-        public override bool Equals(object obj)
-        {
-            if (obj is null)
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, obj))
-            {
-                return true;
-            }
-
-            return obj.GetType() == GetType() && Equals((TesInput)obj);
-        }
-
-        /// <summary>
-        /// Returns true if TesInput instances are equal
-        /// </summary>
-        /// <param name="other">Instance of TesInput to be compared</param>
-        /// <returns>Boolean</returns>
-        public bool Equals(TesInput other)
-        {
-            if (other is null)
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, other))
-            {
-                return true;
-            }
-
-            return
-                (
-                    Name == other.Name ||
-                    Name != null &&
-                    Name.Equals(other.Name)
-                ) &&
-                (
-                    Description == other.Description ||
-                    Description != null &&
-                    Description.Equals(other.Description)
-                ) &&
-                (
-                    Url == other.Url ||
-                    Url != null &&
-                    Url.Equals(other.Url)
-                ) &&
-                (
-                    Path == other.Path ||
-                    Path != null &&
-                    Path.Equals(other.Path)
-                ) &&
-                (
-                    Type == other.Type ||
-                    Type.Equals(other.Type)
-                ) &&
-                (
-                    Content == other.Content ||
-                    Content != null &&
-                    Content.Equals(other.Content)
-                );
-        }
-
-        /// <summary>
-        /// Gets the hash code
-        /// </summary>
-        /// <returns>Hash code</returns>
-        public override int GetHashCode()
-        {
-            unchecked // Overflow is fine, just wrap
-            {
-                var hashCode = 41;
-                // Suitable nullity checks etc, of course :)
-                if (Name != null)
-                {
-                    hashCode = hashCode * 59 + Name.GetHashCode();
-                }
-
-                if (Description != null)
-                {
-                    hashCode = hashCode * 59 + Description.GetHashCode();
-                }
-
-                if (Url != null)
-                {
-                    hashCode = hashCode * 59 + Url.GetHashCode();
-                }
-
-                if (Path != null)
-                {
-                    hashCode = hashCode * 59 + Path.GetHashCode();
-                }
-
-                hashCode = hashCode * 59 + Type.GetHashCode();
-                if (Content != null)
-                {
-                    hashCode = hashCode * 59 + Content.GetHashCode();
-                }
-
-                return hashCode;
-            }
-        }
-
-        #region Operators
-#pragma warning disable 1591
-
-        public static bool operator ==(TesInput left, TesInput right)
-        {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(TesInput left, TesInput right)
-        {
-            return !Equals(left, right);
-        }
-
-#pragma warning restore 1591
-        #endregion Operators
-    }
-}
