+++ src/TesApi.Web/Scheduler.cs
@@ -2,14 +2,15 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Globalization;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using TesApi.Models;
+using Tes.Extensions;
+using Tes.Models;
+using Tes.Repository;
 
 namespace TesApi.Web
 {
